+++ apps/web/src/_pages/console/OrganizationsPage/index.ts
@@ -1,1 +1,2 @@
 export { OrganizationsPage } from './OrganizationsPage'
+export { OrganizationsPageSkeleton } from './OrganizationsPageSkeleton'