+++ apps/web/src/_pages/console/index.ts
@@ -1,3 +1,5 @@
 export { ConsoleLoginPage } from './ConsoleLoginPage'
-export { OrganizationsPage } from './OrganizationsPage'
-export { OverviewPage } from './OverviewPage'
+export { ConsolePageFrame } from './ConsolePageFrame'
+export { OrganizationsPage, OrganizationsPageSkeleton } from './OrganizationsPage'
+export { OverviewPage, OverviewPageSkeleton } from './OverviewPage'
+export { UsersPage, UsersPageSkeleton } from './UsersPage'